requirements.txt
openai
python-dotenv
azure-search-documents==11.4.0
azure-ai-textanalytics
argparse
pandas
tenacity